# Whitelisting for Performance