--- a/edge_test.rb
+++ b/edge_test.rb
@@ -4,10 +4,13 @@ require 'browsermob-proxy'
 require 'capybara'
 require 'capybara/dsl'
 require 'capybara/rspec'
+require 'uri'
 
 class EdgeTest
   COMPARISON_PARAMS = ["title", "url", "date", "pid", "tags", "channels", "authors"]
   attr_accessor :url, :driver, :session
+  EDGE_REGEX = /http:\/\/edge.simplereach.com.*/
+
 
   def initialize(url, options = {})
     @url = url
@@ -17,12 +20,16 @@ class EdgeTest
     @session = Capybara::Session.new(driver)
   end
 
-  def self.run
+  def self.run(url)
+    host = URI.parse(url).host
+    url_to_whitelist = Regexp.new("((http|https):\/\/)?#{host}.*")
+    proxy.whitelist([EDGE_REGEX, url_to_whitelist], 404)
     server.start
     setup_drivers
   end
 
   def self.stop
+    proxy.clear_whitelist
     proxy.close
   end