For URL http://www.thedrive.com/accelerator/4356/lewis-hamilton-will-be-the-greatest-f1-driver-of-all-time

Whitelisting

- Max Memory: 323.4 MB
- Test time: Finished in 1 minute 26.47 seconds (files took 0.51465 seconds to load)

Not Whitelisting

- Max Memory: 573.1 MB
- Test Time: Finished in 1 minute 53.04 seconds (files took 1.11 seconds to load)